+++ policyengine_uk/variables/gov/dwp/AFCS.py
@@ -1,13 +1,13 @@
 from policyengine_uk.model_api import *
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class AFCS(Variable):
     value_type = float
     entity = Person
     label = "Armed Forces Compensation Scheme"
     definition_period = YEAR
     unit = GBP
+    uprating = "calibration.uprating.september_cpi"
 
     def formula(person, period, parameters):
         return person("AFCS_reported", period)
