+++ policyengine_uk/variables/gov/dwp/ESA_income.py
@@ -45,7 +45,6 @@ class claims_ESA_income(Variable):
         return would_claim & claims_legacy_benefits
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class ESA_income(Variable):
     value_type = float
     entity = BenUnit
