+++ policyengine_uk/variables/gov/dwp/council_tax_benefit.py
@@ -10,7 +10,6 @@ class council_tax_benefit_reported(Variable):
     unit = GBP
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class council_tax_benefit(Variable):
     value_type = float
     entity = BenUnit
