+++ policyengine_uk/variables/gov/dwp/student.py
@@ -1,7 +1,6 @@
 from policyengine_uk.model_api import *
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class student_loans(Variable):
     value_type = float
     entity = Person
@@ -10,7 +9,6 @@ class student_loans(Variable):
     unit = GBP
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class adult_ema(Variable):
     label = "Adult EMA"
     documentation = "Educational Maintenance Allowance for adults"
@@ -20,7 +18,6 @@ class adult_ema(Variable):
     unit = GBP
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class child_ema(Variable):
     label = "Child EMA"
     documentation = "Educational Maintenance Allowance for children"
@@ -30,7 +27,6 @@ class child_ema(Variable):
     unit = GBP
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class access_fund(Variable):
     label = "Access Fund"
     documentation = "Access Fund for educational assistance"
@@ -40,7 +36,6 @@ class access_fund(Variable):
     unit = GBP
 
 
-@uprated(by="calibration.uprating.september_cpi")
 class education_grants(Variable):
     label = "Education grants"
     documentation = "Grants for education"
