+++ policyengine_uk/variables/household/wealth/financial.py
@@ -1,9 +1,6 @@
 from policyengine_uk.model_api import *
 
 
-@uprated(
-    by="household.wealth.national_balance_sheet.household.financial_net_worth"
-)
 class gross_financial_wealth(Variable):
     label = "Gross financial wealth"
     entity = Household
@@ -12,9 +9,6 @@ class gross_financial_wealth(Variable):
     unit = GBP
 
 
-@uprated(
-    by="household.wealth.national_balance_sheet.household.financial_net_worth"
-)
 class net_financial_wealth(Variable):
     label = "Net financial wealth"
     entity = Household
