+++ policyengine_uk/variables/input/geography.py
@@ -1,0 +1,24 @@
+from policyengine_uk.model_api import *
+from policyengine_uk.variables.household.demographic.locations import BRMAName
+from policyengine_uk.variables.household.demographic.geography import Region
+
+label = "Geography"
+
+
+class BRMA(Variable):
+    value_type = Enum
+    possible_values = BRMAName
+    default_value = BRMAName.MAIDSTONE
+    entity = Household
+    label = "Broad Rental Market Area"
+    definition_period = YEAR
+
+
+class region(Variable):
+    value_type = Enum
+    possible_values = Region
+    default_value = Region.LONDON
+    entity = Household
+    label = "region"
+    documentation = "Area of the UK"
+    definition_period = YEAR
